Fix non-working file picker in TimezoneExtensionPrefsWidget4
The call to get_root() to set the GtkWindow the file picker should be
transient for in TimezoneExtensionPrefsWidget4 needs to be done over
a GtkWidget, not a GObject, so call it over this.box.

This fixes the file picker to select a custom path for the people.json
file in the preferences dialog.

Closes #59